Cleanup the API for init/fini:
The application calls lt_topo_init(&topo, NULL) and lt_topo_fini(topo)
The allocation of the topo structure is done internally.

lt_set_fsys_root() cannot be public since it has to be called between
the internal allocation and topo_discover.
So drop it from the public API and make it an optional parameter in
lt_topo_init().

This commit was SVN r26.