Only merge levels in lstopo-text when cpuset is...
Only merge levels in lstopo-text when cpuset is equal and arity is 1.

It'll help machines with NUMA nodes without CPUs (we want to keep
these even if their cpuset is empty, but we don't want a single
non-empty node to be merged with the machine).

This commit was SVN r135.