Add cpuset comparison helpers and use them for ...
Add cpuset comparison helpers and use them for topological sorting:
* topo_cpuset_compar_first() compares using the least significant bit,
  so it's similar to comparing topo_cpuset_first() values
  (except that empty cpusets are considered higher than anything)
* topo_cpuset_compar() compares using the most significant bits
  (empty cpusets are considered lower than anything).

It's not clear to me which one is better for topological sorting.
They usually mean the same, except when the topology isn't symmetric
(for instance in 16em64t-4d2c2t-offlines). I am using the former
for now so that test topology outputs don't need to be changed.

This commit was SVN r433.