Add prev/next_cousin pointers in topo_obj. This...
Add prev/next_cousin pointers in topo_obj. This is the list of all objects
from same type in the machine. prev/next_sibling is the list of all children
of the same father. So the latter is smaller than the former.

Fix synthetic topo and a macros which wrongly assumed the existing
next/prev_sibling pointers were actually our new prev/next_cousin.

This commit was SVN r436.