Rework topo_get_type_or_below/above_depth() so ...
Rework topo_get_type_or_below/above_depth() so that we report an actual
depth instead of TOPO_TYPE_DEPTH_MULTIPLE.
For below, return the depth of the highest existing level with order
at least as the given type.
For above, return the depth of the lowest existing level with order
at most as the given type.

This commit was SVN r617.