Rename pkginclude_HEADERS into DX_INPUTS so tha...
Rename pkginclude_HEADERS into DX_INPUTS so that the doc directory
doesn't install all doxygen input files inside include/libtopology/
while the include directory already installs the right headers
in include/.

pkginclude_HEADERS could be used if the list of headers to install
and the list of doxygen input files was the same, and if there was
no install rules in include/Makefile.am.
But DX_INPUTS contains some non-headers, and some headers are not
in DX_INPUTS.

This commit was SVN r717.