Add the TOPO_FAKE environment variable to manua...
Add the TOPO_FAKE environment variable to manually enable/disable
the is_fake flag of the topology.
May be used to load a XML topology instead of reparsing the whole
local topology, without setting the is_fake flag.

Use it in gather-topology.sh to properly get the topology output
as if it were fake.

This commit was SVN r807.