topobind: use topomask helpers so that we may n...
topobind: use topomask helpers so that we may now directly bind using
a space-separated list of objects (eveything that topomask supports).

Now that topobind is clever, topomask is useless is most basic cases.
topomask is now useful when manipulating objects/cpumasks without
binding immediately. Or when we need to compose multiple invokations
to enfore cpumask operation priorities.

This commit was SVN r862.