Fixes #3.
Have fixed versions of the man pages for the executables in svn under
the .1in filename pattern.  These files are processed during "make
all" to make the corresponding .1 files.  The changes that are added
in are the project name, date, and release version.  This gives
nice/proper headers and footers on the man pages that are specific to
the version that they came from.

This also fixes the issue of having the man pages depend on the
executables, which was causing problems for making nightly tarballs.

This commit was SVN r1001.