Add a ibverbs test which prints the cpuset of e...
Add a ibverbs test which prints the cpuset of each infiniband device.

It uses ibverbs so we have to check for its headers and lib
at configure. I wonder if we should just make this an example
and let the user build/link it properly when testing it.

This commit was SVN r1013.