Use the same PCI_DEVICE type for all devices (e...
Use the same PCI_DEVICE type for all devices (except bridges).
We'll let the above layers (or helpers) sort out the interesting ones
(or we'll provide filtering methods with topology flags)

Add PCI device and bridge object attributes and start filling
them for real (even if we do nothing with objects for now).

This commit was SVN r1036.