Insert all PCI Devices and Bridges in a tree of...
Insert all PCI Devices and Bridges in a tree of hwloc object
according to their bus ids.

Only the first/last_child, prev/next_siblings and arity fields
are set so far.

Print the hierarchy at the end instead of inserting it in the
main topology.

This commit was SVN r1038.