Take all PCI objects and place them under new m...
Take all PCI objects and place them under new manually-created
PCIBridges that correspond to the machine hostbridges between
the main bus and the upstream PCI busses.
Each individual domain/bus gets its own hostbridge.

Compute the cpuset of nearby cpus for each hostbridge on Linux
but it's not used yet (we still attach to the top system object).

This commit was SVN r1054.