Use AC_MSG_CHECKING/AC_MSG_RESULT so that we do...
Use AC_MSG_CHECKING/AC_MSG_RESULT so that we don't have naked echo's
in configure (and therefore the messages show up in config.log).  Also
print a Big Hairy Warning (that also shows up in config.log) if we are
on an unsupported platform.

This commit was SVN r1224.