Linux: Don't write the whole set of cpuset-disa...
Linux: Don't write the whole set of cpuset-disabled cpus directly
in the offline_cpuset since we don't want unexisting CPUs there.

Wait for the main sysfs parsing loop to filter the existing ones
before actually the offline ones in offline_cpuset.

We don't support the offline_cpuset when /proc/cpuinfo is used
instead of /sys.

This commit was SVN r1328.