Always print an error message if execvp fails (...
Always print an error message if execvp fails (vs. only displaying
when -v is specified).  Also print the argv[0] that we tried to
execvp, since perror won't (e.g., if the user accidentally mis-typed
an option and hwloc-bind tried to use that as the argv[0]).

This commit was SVN r1429.