Add routines to read the list of thread tids of...
Add routines to read the list of thread tids of a process and apply
a callback to all of them. Then, re-read the list again and try again
if it changed in the meantime.

Use this to implement hwloc_linux_set/get_pid_cpubind().

Use them in hwloc_linux_set/get_proc_cpubind() when PROC is given.

This commit was SVN r1542.