Don't enforce a SYSTEM object at the root of th...
Don't enforce a SYSTEM object at the root of the tree.
Use MACHINE by default, and change it into SYSTEM if kerrighed is detected
or if a synthetic topology with machine levels is used.

Update all linux test outputs according.

Update lstopo to not show the index of the top level instead
of a system level.

Make check passes.

Refs #29.

This commit was SVN r1626.