Read /proc/mounts to detect the cpuset/cgroup m...
Read /proc/mounts to detect the cpuset/cgroup mount points instead
of testing /cpuset /dev/cpuset /cgroup and /dev/cgroup.

The cgroup package seems to mount some stuff on /mnt/cgroups/*.
And the default cgroup mount point was still being discussed recently
on LKML. So it's better to not hardwire anything in hwloc.

This commit was SVN r1749.