Fix support for Linux kernels with CONFIG_NR_CP...
Fix support for Linux kernels with CONFIG_NR_CPUS not a multiple of 32.

If the cpumask starts with a bitmask with less than 8 digits,
fgets would return missaligned submasks, and get screwed.
Just use fscanf instead.

Thanks to Andreas Herrmann for reporting the problem and testing the patch
(was on a 48 core AMD machine with CONFIG_NR_CPUS=48).

This commit was SVN r1755.