Don't use a random exponential loop to read the...
Don't use a random exponential loop to read the list of tids in hwloc_linux_get_proc_tids

Get a good estimation of the number of tids with with stat.st_nlinks
and then expand the array by steps of 8 if needed.

Thanks to Bert Wesarg for the patch.

This commit was SVN r1823.