When retrying the loop during a get_proc_cpubin...
When retrying the loop during a get_proc_cpubind on Linux,
clear the cpuset before readding each thread cpuset.

We add the array 'index' to the callback and use it to clear
the current cpuset when needed, and also to easierly detect
the first thread.

Thanks to Bert Wesart for the report and patch,
with major changes from me.

This commit was SVN r1838.