Drop HWLOC_CPUSUBSET_SUBSET and use set->ulongs...
Drop HWLOC_CPUSUBSET_SUBSET and use set->ulongs[i] directly when
the caller knows that i is a valid index within this cpuset array.

In the general case, when we don't know if i is valid, use a new
HWLOC_CPUSUBSET_READULONG to read the ulong or generate it from
the infinite flag.

This commit was SVN r1991.