Deoptimize many cpuset routines by using HWLOC_...
Deoptimize many cpuset routines by using HWLOC_CPUSUBSET_READULONG()
all the time instead of manually handling the cases where a ulong index
is valid in one of the input cpuset and/or the other.

The performance gain wasn't so huge (0-20%), and the code was waaaaay
harder to read/maintain. And things should be much easier to deal
with when we'll add sparse cpuset support.

This commit was SVN r2107.