When mounting a cpuset mntpnt with on 2.6.24+, ...
When mounting a cpuset mntpnt with on 2.6.24+, a cgroup mntpnt is
actually created in /proc/mounts.
For backward compatibility, the "noprefix" mount option is added
so as to have "cpus" and "mems" filenames (instead of the usual
"cpuset.cpus" and "cpuset.mems" in cgroup mntpnt).
Fix the cgroup/cpuset detection code to handle this properly by
looking for the "noprefix" option.

Thanks to Michael Raymond for the report.

This commit was SVN r2355.