Put distance matrices under objects instead of ...
Put distance matrices under objects instead of in the topology

This way, we may place different matrices under different objects
if we only know a subpart of the distances (for instance when
agregating different machines).

The distance matrix is now stored with the corresponding relative
depth in a public array in object attributes.

This commit was SVN r2464.