Add CUDA helpers to retrieve the locality of CU...
Add CUDA helpers to retrieve the locality of CUDA devices

One for the Driver API (cuda) and one for the Runtime API (cudart).

This requires the upcoming cuda 3.2.

No configure-stuff yet since cuda doesn't seem to have proper pkg-config stuff and so on.
Use command-lines like below to build the tests:
  make cudart -C tests/ CFLAGS=-I/usr/local/cuda/include LDFLAGS="-lcuda -lcudart -lhwloc -L/usr/local/cuda/lib64/"
  make cuda -C tests/ CFLAGS=-I/usr/local/cuda/include LDFLAGS="-lcuda -lhwloc"

This commit was SVN r2483.