Add *.exclude support to tests/linux/test-topol...
Add *.exclude support to tests/linux/test-topology.sh to ignore part of tarballs

If foo.exclude exists, it is passed to tar --exclude-from so as to not extract
some files of the given tarball.

Use .source+.exclude instead of duplicating/modifying the huge 256ppc-8n8s4t.tar.gz
Saves about 300kB for make dist

This commit was SVN r2516.