Do not call xmlCleanupParser at all
The documentation says that it frees global variables in the libxml library
and should be used with caution. We don't know if some other calls, either
simultaneous or later, will reuse those variables.

That's a leak of 926 bytes on my machine, per process, nothing horrible,
but it's safer.

This commit was SVN r2562.