document where we got the distance matrix from ...
document where we got the distance matrix from and add some questions

Linux on x86, ia64, sparc and powerpc seems to all conform to the ACPI
SLIT spec (but sparc doesn't seem to report it to user-space).

Linux on mips/sgi-ip27 (looks like it may be a SGI Origin200/2000)
might be doing something different (local=0, same router=1, and more
(100+) for others)

This commit was SVN r2582.