Rework the distance API for normalized memory l...
Rework the distance API for normalized memory latency

Linux upto 2.6.36 reports ACPI SLIT-like distances in sysfs, which means
they are memory latencies. Rename the fields of the distance structure
accordingly. Normalize values to get 1.0 on the diagonal (and store
the normalization factor on the side).

When the minimal distance is 0, ignore the matrix since it is likely a
number of hops instead of a latency.
The only known case so far is SGI IP27 (Origin 200/2000 with MIPS processors)
where Linux reports the number of router hops between nodes.
This (and HT/QPI links) could be stored as another array of unsigned values
in the distance structure. This way, we'd have a distance structure for each
depth, containing either latencies or hops (or both).

hwloc/helper.h still needs some improvements according to all this.

This commit was SVN r2589.