Don't use dirent->d_type since some FS do not s...
Don't use dirent->d_type since some FS do not support it

Use stat.st_mode instead.

Fixes the device-tree code when running from a Linux tarball
on some non-trivial filesystems.

Thanks a lot to Christopher Samuel for reporting and debugging
the problem, and for testing the patch.

This commit was SVN r2799.