Add HWLOC_TOPOLOGY_FLAG_NO_PCI_BRIDGE to only s...
Add HWLOC_TOPOLOGY_FLAG_NO_PCI_BRIDGE to only show PCI devices
All PCI bridges are ignored. The host bridge is not ignored since
it is an easy way to know when we enter an I/O subsystem
(you only have to look for a bridge, not for possibly different
types of objects).

1) NO_PCI shows nothing
2) NO_PCI_BRIDGE only shows PCI devices and host bridges
3) default (no flag) shows "interesting" devices (gpu+net+ib+block)
   and bridges that have some children
4) WHOLE_PCI: shows everything

I think we should drop (3) in favor of (2). People will likely not
care about bridges (aka (2) is enough), and those that care can handle
the whole PCI stuff (aka (4) is ok).

And I wonder if we should make (1) the default for "backward compatibility".

This commit was SVN r2875.