Remove C99 notation from hwloc.  This SVN commi...
Remove C99 notation from hwloc.  This SVN commit is the result of a
bunch of work in a branch mercurial repo discussed on the hwloc devel
mailing list in a few different threads, such as this one:

    http://www.open-mpi.org/community/lists/hwloc-devel/2011/02/1948.php

There were a few other minor updates bundled in this commit (e.g.,
removing commented-out code), but removing C99 notation is the biggest
change.

This commit was SVN r3200.