Rework the behavior of hwloc_topology_insert_mi...
Rework the behavior of hwloc_topology_insert_misc_object_by_cpuset/parent

When inserting by cpuset, we may add a new intermediate level in the tree.
Setup all cpusets in this case, so that we don't create a special Misc case
with a cpuset but no complete/online/allowed_cpuset.

When inserting by parent, we don't specify the cpuset of the new object.
So we don't know if it should cover the existing children or be an additional
child. The former can be done by inserting by cpuset. So we do the latter:
we append a new children with no cpusets (it can't overlap with other
children).

Clarify all this in the doc, and make it clear that inserting by parent
is useful for annotating the topology, not for changing the topology tree.

Remove the now useless Misc ugly hack in the restrict code: Misc object
now have all cpuset fields defined, or none, as any other object.

This is an intrusive behavior change but I don't expect many people
(apart from lstopo --ps) to insert Misc objects (and rely on the actual
values of the object cpusets).

This commit was SVN r3299.