When attaching hostbridges to the topology, che...
When attaching hostbridges to the topology, check that we found
a parent with the exactly required cpuset.
If not, insert a intermediate Group object.

This is useful for attaching IOHubs on Nehalem/Westmere-EX machines
with more than 2 sockets because their IOHub is connected to two
sockets only. Without this, the IOHub would be attached to the
parent, for instance the entire machine in case of quad-socket
machines.

This commit was SVN r3344.