Fix an invalid memory read during distance rest...
Fix an invalid memory read during distance restrict

We were reading an additional line.
It never actually caused any corruption because the destination
is valid and not read anymore (we're reducing the matrix size
without reallocating it smaller).

Thanks to valgrind.

(needs backport in v1.2 and v1.3)

This commit was SVN r3652.