Create a group object above a set of objects co...
Create a group object above a set of objects connected through a distance matrix

This new object is usually ignored.
In the case of submatrices, the list of considered objects may not have
an exact ancestor. We add it so that we can store the matrix distance there.

This commit was SVN r3717.