Temporarily disable commit r3717 (adding a grou...
Temporarily disable commit r3717 (adding a group exactly above distances matrices)

It breaks 128ia64-17n4s2c linux test topology because we fails to merge
the newly added group with the machine object. The reason is that the
last numanode has no CPU, hence the insert code considers it as not
comparable with the group. We'll fix this by comparing nodeset too.

This commit was SVN r3724.

The following SVN revision numbers were found above:
  r3717 --> 1c081576dc95fd3520c3e258f7980b00efda170b