When we insert a group object to cover a distan...
When we insert a group object to cover a distance matrix,
it must cover the cpusets and the nodesets of all objects
connected by the matrix. Otherwise, cases with CPUless NUMAnode
can end up storing the matrix in strange objects (with negative
relative depth).

tests/linux/32amd64-4s2n4c-cgroup is an example, fixed here

This commit was SVN r3774.