Add notes to encourage people (including me!) t...
Add notes to encourage people (including me!) to go a) go read the
formal docs; don't just rely on reading hwloc.h, and b) check the
Terms and Defs section of the docs for a wonderful picture showing
child/sibling/cousin relationships, and an example of an asymmetric
topology. 

This commit was SVN r3853.