Ignore distance matrices in case of multi-node ...
Ignore distance matrices in case of multi-node topologies

The current code relies on cpuset for grouping and for finding
the root object where the matrix is attached, this can't work
with multinode because some objects have no cpusets.

Unfortunately, at least one user already reported his interest
in having distances between multiple nodes. We'll see what we
can do about it.
One idea could consist in adding a hostset on the side of cpusets
and nodesets to identify which hosts are included in a object.

This commit was SVN r3916.