Before v1.3, inserting distance matrices in the...
Before v1.3, inserting distance matrices in the topology can get
wrong if we have CPU-less NUMA-nodes because we only look at cpusets
(not at nodesets). We end up inserting the topology in the wrong
common ancestors, which means the relative_depth can get wrong.

This patch ignores the matrix if the relative_depth isn't <= 0.
This is a sanity check, but it also ensures that we won't
export XML that cannot be imported (relative_depth must be != 0).

Thanks to Chris Yeoh from IBM for reporting the problem.
(http://www.open-mpi.org/community/lists/hwloc-devel/2011/11/2524.php).

This commit was SVN r3951.