Forward port v1.2-ompi commit r3951 to trunk.
It's less important here because we properly insert distance matrices
by looking at both cpusets and nodesets. But it's still a good sanity
check to make sure that relative_depth is >= 0. And it ensures that
we won't export XML that cannot be imported (relative_depth must be != 0).

This commit was SVN r3952.

The following SVN revision numbers were found above:
  r3951 --> 5f0734c527bd29b8478e7639c66088150209ff03