Rework the solaris code that gathers CPUModel/T...
Rework the solaris code that gathers CPUModel/Type (commit r3964)

Terry's code stored the attributes in the machine object and assumed
that all sockets are identical (should be true for all platforms
that matter for Oracle and OMPI). Move these attributes to the socket
object to match what Linux does.

We still assume that all sockets are identical because reworking
the chiptype detection code to handle different sockets isn't trivial.

We could keep a copy of the attributes in the machine object when
all sockets are identical, but we'd have to document it and users
may start asking why we don't have copies on other object levels.

This commit was SVN r4008.

The following SVN revision numbers were found above:
  r3964 --> 65e8d420c548a8e67ff6cc780012ed7ce4b863b2