lstopo --horiz/--vert now accepts =type1,type2,...
lstopo --horiz/--vert now accepts =type1,type2,...

This lets you precisely place objects, which is very useful
when you want something practical to put on a website, slide, etc.

Each type specifies a container, i.e. the layout of children
object *inside* this object. It does not specify the layout of
objects of the given types.

This commit was SVN r4027.