Skip linux-libnuma test when the kernel does no...
Skip linux-libnuma test when the kernel does not support numa

So far we assumed that libnuma behaved as if there was a single
numa node when the machine wasn't numa or when the kernel wasn't
numa aware.
That doesn't work anymore since libnuma 2.0.6 because numa_all_nodes
does not get properly initialized (at least when linking dynamically)
and the kernel doesn't have CONFIG_NUMA.
The doc actually says that libnuma behavior is undefined in this
case, so don't try to make sure we're compatible in this case, just
skip the entire test.

Document this in the helper header.

This commit was SVN r4046.