Fix support for sched_set/getaffinity() on redh...
Fix support for sched_set/getaffinity() on redhat 8

Their very old prototypes use a length and ulong* instead of cpu_set*.
We can't assume that this one will never appear in sched.h.
At leasdt redhat 8 defines it.

So we have to check whether it's in the header before redefining it
with syscall3.

This commit was SVN r4059.