Fix the documentation of hwloc_pid_t and hwloc_...
Fix the documentation of hwloc_pid_t and hwloc_thread_t

Don't define them to pid_t and pthread_t in doxygen, because
the website doc would be wrong for windows.

But still define hwloc_thread_t to itself so that doxygen
does not ignore the #ifdef around the relevant API.

Also make sure all functions that use one of these types
properly documents what it is on Unix and Windows.

We can't easily have dedicated docs for these types in doxygen
because they are #define in autogen/config.h at configure.

This commit was SVN r4081.