After some off-list discussion, put the v1.3 se...
After some off-list discussion, put the v1.3 series back to Libtool's
--version-number scheme, and simply bump the .so VERSION to 4:2:4.
The rationale here is that the versions in v1.3 and v1.3.1 weren't
what we intended (we meant to use --version-info, not
--version-number).  But since they're broken already, just leave the
entire v1.3 series using --version-number and leave the major .so
version number identical (4) so that ABI compatibility is ensured.

We'll fix .so versioning for real starting with v1.4.1.

This commit was SVN r4278.