Backport something kind of similar to trunk com...
Backport something kind of similar to trunk commit r4244 in v1.3 branch:
* Fix assertion on distance group cpuset/nodeset
  When merging is disabled, some objects involved in distance matrices may disappear.
  The distance group object may thus loose children and appear to have too large
  cpusets and nodesets. Nothing bad, just fix the assertions.

This patch wasn't backported earlier because the target code is very
different in trunk/v1.4. The assertions are even worse in v1.3 because
we can get a NULL root object there. Ignore the distance matrix in
this case as well.

Thanks to Stephanie Moreaud for reporting the problem.

This commit was SVN r4314.

The following SVN revision numbers were found above:
  r4244 --> aee0523d98665e7717cf6f67fd920d0e4fe0e399