Replace an assertion with a error messages in t...
Replace an assertion with a error messages in the distance code

If we ever find a NULL covering object instead of a proper root
object for a distance matrix, no need to break with an assert,
just ignore the matrix and print an error message.
It should not happen, but this code is tricky, especially when
there are cgroups and ignoring options, so I'd rather workaround
the problem (and still have people report it) instead of just
asserting false.

This commit was SVN r4315.