Add Port info attribute to network OS devices i...
Add Port info attribute to network OS devices inside OpenFabrics PCI
devices so as to identify which interface corresponds to which port.

Don't do it for all network OS devices because the dev_id sysfs
attribute for OFED (dev_id = port-1) is not strictly equivalent
to its generic meaning (shared nic id, 0 is most cases).

This commit was SVN r4417.