Fix the retry logic when working on all threads...
Fix the retry logic when working on all threads of a process

Some callbacks may fail if a thread disappears in the meantime.
If some fail, but not all, try again.
If all failed, but the tid list didn't change, report the last error.

The only case where we could report an error is when all thread
disappear in the meantime and are replaced with as many threads
with the same tids. Very unlikely :)

Thanks to Vlad Roubtsov for the report and testing the patches.

This commit was SVN r4443.