Fix support for more than 32 processors on Wind...
Fix support for more than 32 processors on Windows

ulong is always 32bits there, so bitmap manipulations were wrong.
We use sizeof(void*) to handle the 32/64 cases.

This commit fixes #70.

Thanks a lot to Hartmut Kaiser for reporting this issue and
testing patches.

I'll backport this to v1.4 after some feedback.

If things work, we should likely add a windows-specific
hwloc_bitmap_set_ith_ULONG_PTR() to factorize all #if added here.

The 32bit version of the this code may still be broken because
windows32 affinity support is broken as explained in Remarks on
  http://msdn.microsoft.com/en-us/library/windows/desktop/dd405488%28v=vs.85%29.aspx
We may need some workarounds. Hopefully win32 machines with more
than 32 procs are rare.

This commit was SVN r4460.