Only create a logical distance matrix if the ro...
Only create a logical distance matrix if the root object covers the right numbers of objects

This check may fail if the user adds a distance matrix that contradicts the topology,
we fail to insert the root (intersection without inclusion), and later we cannot
insert the logical matrix.

This commit was SVN r4715.