nolibxml/export: Store the number of children a...
nolibxml/export: Store the number of children and content existency in the state

This fixes the layout of objects with respect to userdata content.
We don't know in advance whether the userdata export() callback
will actually export something. So we don't know for sure if the
object XML node will have some userdata XML children nodes.

This fixes tests/hwloc_object_userdata when using nolibxml.

This commit was SVN r4790.