Add --quiet to hwloc-calc and hwloc-bind
This hides non-fatal error messages such as non-existing locations.
We don't abort on such errors because it's convenient to let the
user specify ranges that are too large, for instance when doing
things like socket:all.core:3-4 on asymmetric topologies.

For hwloc-bind, it's useful when combined with --force.

This commit was SVN r4837.