Fix the management of the thissystem flag
It is initialized to 1 during init().
If the application calls set_foo() to change the backend,
it may overwrite the flag with set_flags().
If the application changes the backend with env vars,
it may overwrite the flag with HWLOC_THISSYSTEM env var.

Things must be processed in this order, and the flag must
be ready early during the discovery because part of the
code uses it.

HWLOC_FORCE_THISSYSTEM was not stronger than the correct
HWLOC_THISSYSTEM env var, so it's removed
(it was never documented anyway).

The main problem that this commit fixes is that set_flags()
was processed to late and would overwrite what the env var
did. But the env vars are stronger that set_*().
Otherwise, it's mostly cleanups, reordering and documentation.

This commit was SVN r4873.