Fix invalid command-line parsing errors in hwlo...
Fix invalid command-line parsing errors in hwloc-distrib and hwloc-distances

Options that need an argument must abort if argc<2, not if argc<=2.
The bug only appeared at the very end of the line, that's why we
didn't notice it, but it has been here for a long time.

This commit was SVN r4900.