Fix doc install when standalone but not buildin...
Fix doc install when standalone but not building the doc

We still need to enter doc/ during make install (if standalone),
even if building doc is disabled.

So always enter doc/ (if standalone) and check in there what to
do depending on whether HWLOC_BUILD_DOXYGEN and/or HWLOC_INSTALL_DOXYGEN
are set.

Ideally, we would only enter doc/ if HWLOC_BUILD_DOXYGEN,
or doing if make install with HWLOC_INSTALL_DOXYGEN,
but that's not easy to do.

This reverts:
r4676:
  Stop protecting things with HWLOC_BUILD_DOXYGEN (almost)
r4686:
  We don't enter doc/ unless standalone, but the reverse isn't true
  (doc can be disabled in standalone). So make sure we only enter
  doc/ when standalone and doc is enabled.

We're back to what we did in v1.5.

This commit was SVN r5045.

The following SVN revision numbers were found above:
  r4676 --> dcabb1acd77b783d2cfb037ea0752cb0e84dbb15
  r4686 --> 5389d2105a9c7342a129aeefeff6c4c23d2379f4