Update hwloc_topology_set_*() documentation wit...
Update hwloc_topology_set_*() documentation with respect to (not) clearing the topology

The component branch changed this behavior from
  clear the topology and try to enable a new backend (which may fail)
to
  try to create a new backend, and (on success) replace the previous one,
  while the topology isn't touched

The application-visible behavior (after load()) changed if set_foo() succeeds
before set_bar() fails before load(). Fortunately, applications are not expected
to play with this, so changing this behavior is OK.

This commit was SVN r5057.