OpenCL: Disable the entire plugin if the AMD ex...
OpenCL: Disable the entire plugin if the AMD extension isn't found

It works fine (does nothing) but it's better to disable the entire
plugin for now to avoid confusing the user.

By the way, also abort if --enable-opencl, found the lib, but didn't
find the extension.

This commit was SVN r5083.