Move component symbols declarations from compon...
Move component symbols declarations from components.h to static-components.h

We don't want to mix plugins-symbols-for-core and core-symbols-for-plugins
in the same header file.

The former are only needed when building components statically, and only
used through the configure-generated array in static-components.h.
So declare them there too.
By the way, this avoids declaring all possible components symbols while
those built as plugins are not actually available.

Refs #88

This commit was SVN r5089.