Add a NVML component/plugin
This is useful for batch schedulers that want to get the locality
of NVIDIA GPUs without actually using CUDA directly.

We add the correponding nvml%d OS devices.

Note that the NVML ordering may be different from the CUDA ordering
(looks like CUDA reorders based on performance).

Note that libnvidia-ml.so is distributed within CUDA
while nvml.h is only available in the Tesla Development Kit.

Refs #71

This commit was SVN r5139.