x86: Fix crashes in case of admin-restriction s...
x86: Fix crashes in case of admin-restriction such as Linux cgroups/cpusets

Initialize the info structure before trying to bind,
so that we don't try to summarize random data later
if binding failed because of cgroups.

Never noticed before v1.6 because the x86 backend was rarely used.
But it's used on Linux after the linux backend by default now.

This commit was SVN r5157.