pci: don't use libpci/pciutils unless --enable-...
pci: don't use libpci/pciutils unless --enable-libpci is given

If libpciaccess is available, we just use it.
If not, if libpci is available, we warn the user that he could
enable it but we didn't by default because of the license.
If --enable-libpci is passed, libpci is used, even when libpciaccess is available.

This commit was SVN r5306.