Merge the GL branch into trunk
I am not merging with svn merge because the history is horrible.

This adds GPU OS devices such as ":0.0" for each NVIDIA display,
using the NVIDIA X-CONTROL X11 extension through the NVCtrl library.
It may later be extended to windows with WGL_NV_gpu_affinity and
WGL_AMD_gpu_association, or to OS X with CGL renderer IDs.
(that's why it's called GL and not just X11 display).

We also have a hwloc/gl.h to convert between :X.Y (string or pair)
and OS device objects.

This commit fixes #54.

Thanks to Marwan Abdellah and Stefan Eilemann (Blue Brain Project, EPFL)
for requesting, writing and testing some code.

This commit was SVN r5358.