Backport libpciaccess in v1.4 branch (trunk commits 5332,5340,5382)
v1.4.3rc1 will contain this big one and several other fixes that
were queued in the branch a while ago. NEWS documents clearly that
the release misses many other fixes that are only in v1.5 or
later, and that v1.4.3 may only be used when upgrading to v1.5
or later is not possible.

This commit was SVN r5390.