tools: don't restrict I/O options in usage() on...
tools: don't restrict I/O options in usage() on libpci being enabled

I/O options (--no-io, etc) are of course available with libpciaccess.
They are actually always available, but they may be useless
if there's no I/O component/plugin or if it finds nothing, but
that hard to detect are runtime.
So just always have them in usage()
(they are always in the manpage anyway).

This commit was SVN r5412.