hwloc.h: improve doxygen syntax for long member...
hwloc.h: improve doxygen syntax for long member name/documentation

Use /** before the field rather than /**< after the field when
the field name or description is very long.
Placing /**< on the right the field causes many linebreaks. And
placing /** on the previous line is better than /**< on the next line.

This makes documentation of restrict flags, cpubind flags, membind
flags and policy much easier to read.

This commit was SVN r5443.