components: queue plugins in the order we get t...
components: queue plugins in the order we get them (instead of reversed)

lt_dlforeachfile() reads the plugin directory in (alphabetic) order, keep the plugins in the same order.
(if there are multiple directories in the envvar, each one is read in order after the previous one)

This commit was SVN r5506.