configure: Add --with-hwloc-plugins-path to spe...
configure: Add --with-hwloc-plugins-path to specify the install/load directories for plugins.

It's a colon-separated list of directories.
Default is "$prefix/lib/hwloc".

Plugins will be installed in the first directory.
They will be loaded from all of them, in order.

This may be used by packagers to install plugins with different ABIs
in different directories.

This commit was SVN r5541.