distances: remove an overzealous assertion that...
distances: remove an overzealous assertion that breaks when loading an already-grouped-by-distances XML

When grouping by distance, a new group can be merged on insert
if importing from a XML file that contains distance matrices and groups.

Thanks to Laercio Lima Pilla for reporting the crash.

This commit was SVN r5554.