core: warn if insert_by_parent() is out-of-orde...
core: warn if insert_by_parent() is out-of-order

Manual modifications of a XML file can lead to topologies that may
or may not work well. The core assumes children are in order in some places.
HWLOC_DEBUG_CHECK and --enable-debug assert this.

Check this in insert_by_parent() to warn on invalid XML.
I don't think any other backend can cause the same problem so far,
but we'll warn too if it ever occurs.

We cannot easily return an error from there and fix the XML import
to handle it (ignore children, etc.).

This commit fixes #102

This commit was SVN r5574.