Fix the merging of groups when they are I/O obj...
Fix the merging of groups when they are I/O objects

I/O objects must be ignored merge_useless_child(). Two objects
are still identical/mergeable when they don't have the same I/O
children (I/O object locality is a cpuset, it's not attached to
one specific object among identical ones).

This fixes a spurious group object appearing between the machine
and NUMA nodes when importing from XML. The reason is that I/O
objects get added early during XML import, and distance processing
creates a group later. This group must be merged with the machine,
even if the machine has I/O children that the group doesn't have.

This doesn't occur in the non-XML case because I/O are added after
distance groups are created and merged.

This commit also fixes XML tests that contained the above spurious
groups. They shouldn't be in the input/output files at all anymore.

Thanks for Elena Elkina for reporting the problem and testing the patch.

This commit was SVN r5592.