configure: Fix sysctl and sysctlbyname checks
Do a full link test instead of just using AC_CHECK_FUNCS, which just
checks to see if the symbol exists or not.  For example, the prototype
of sysctl uses u_int, which on some platforms (such as FreeBSD) is
only defined under __BSD_VISIBLE, __USE_BSD or other similar
definitions.  So while the symbols "sysctl" and "sysctlbyname" might
still be available in libc (which autoconf checks for), they might not
be actually usable.

Thanks to Pavan Balaji.

This commit was SVN r5598.