configure: Improve _HWLOC_CHECK_DECL
The check for whether an explicitly function declaration is needed
was relying on whether a redefinition of the function throws an error.

This only works if the function declaration is already present in one
of the headers.  If such a declaration is not present, the test might
fail with "implicit function declaration" with the right CFLAGS.  This
leads the m4 macro to think that the declaration is already there in
one of the headers and an additional declaration is not needed.

This commit fixes this by adding a dummy function declaration,
together with the dummy function definition.

This is required for proper fls flsl detection under strict build flags.

Thanks to Pavan Balaji.

This commit was SVN r5599.