Stop embedding libltdl, always use the system-w...
Stop embedding libltdl, always use the system-wide libltdl

The ltdl embedding caused problems to the hwloc embedding such as
  http://www.open-mpi.org/community/lists/hwloc-devel/2013/04/3659.php
We fixed the embedding AM_CONDITIONAL problem in
  https://svn.open-mpi.org/trac/hwloc/changeset/5605
but the generated tarballs now (sometimes) miss libltdl,
causing configure to break.
The patch in the first link above worked around that issue but...

* Embedding ltdl is useful when you rely on recent ltdl features,
  while ltdl 1.5 (couldn't test earlier) looks OK for hwloc,
  and that version is very old and available everywhere.
* the ltdl embedding ability isn't perfect in "recursive" mode
  (we had a hack for its config.h file in our configure
   see http://lists.gnu.org/archive/html/libtool/2012-08/msg00016.html)
* we only need ltdl when --enable-plugins (not by default)

That's enough reasons to consider not embedding ltdl anymore.
We now require people to install libltdl-dev/libtool-ltdl-dev
if they want plugins.

This commit was SVN r5618.